Move AlbumManager to it's own file
